TIP-863: move Category to Enrichment
