+++ src/Pim/Bundle/PdfGeneratorBundle/spec/Renderer/ProductPdfRendererSpec.php
@@ -7,7 +7,7 @@ use Liip\ImagineBundle\Imagine\Cache\CacheManager;
 use Liip\ImagineBundle\Imagine\Data\DataManager;
 use Liip\ImagineBundle\Imagine\Filter\FilterManager;
 use PhpSpec\ObjectBehavior;
-use Pim\Bundle\CatalogBundle\Entity\Category;
+use Akeneo\Pim\Enrichment\Component\Category\Model\Category;
 use Pim\Bundle\PdfGeneratorBundle\Builder\PdfBuilderInterface;
 use Akeneo\Pim\Structure\Component\Model\AttributeGroupInterface;
 use Akeneo\Pim\Structure\Component\Model\AttributeInterface;
